+++ src/main/kotlin/io/github/hellocuriosity/ModelForgeException.kt
@@ -1,3 +1,5 @@
+package io.github.hellocuriosity
+
 import java.lang.RuntimeException
 
 class ModelForgeException(message: String) : RuntimeException(message)
