+++ src/main/kotlin/io/github/hellocuriosity/providers/DoubleProvider.kt
@@ -1,17 +1,18 @@
-package providers
+package io.github.hellocuriosity.providers
 
 import kotlin.random.Random
 
-class FloatProvider(
+class DoubleProvider(
     private val min: Double = DEFAULT_MIN,
     private val max: Double = DEFAULT_MAX,
     private val random: Random = Random
-) : Provider<Float> {
+) : Provider<Double> {
+
     companion object {
         const val DEFAULT_MIN = 1.0
         const val DEFAULT_MAX = 100000.0
     }
 
-    override fun get(): Float =
-        random.nextDouble(min, max).toFloat()
+    override fun get(): Double =
+        random.nextDouble(min, max)
 }
