+++ src/main/kotlin/io/github/hellocuriosity/providers/FloatProvider.kt
@@ -1,18 +1,17 @@
-package providers
+package io.github.hellocuriosity.providers
 
 import kotlin.random.Random
 
-class DoubleProvider(
+class FloatProvider(
     private val min: Double = DEFAULT_MIN,
     private val max: Double = DEFAULT_MAX,
     private val random: Random = Random
-) : Provider<Double> {
-
+) : Provider<Float> {
     companion object {
         const val DEFAULT_MIN = 1.0
         const val DEFAULT_MAX = 100000.0
     }
 
-    override fun get(): Double =
-        random.nextDouble(min, max)
+    override fun get(): Float =
+        random.nextDouble(min, max).toFloat()
 }
