+++ src/main/kotlin/io/github/hellocuriosity/providers/LongProvider.kt
@@ -1,4 +1,4 @@
-package providers
+package io.github.hellocuriosity.providers
 
 import kotlin.random.Random
