+++ src/main/kotlin/io/github/hellocuriosity/providers/Provider.kt
@@ -1,4 +1,4 @@
-package providers
+package io.github.hellocuriosity.providers
 
 interface Provider<T> {
     fun get(): T
