+++ src/main/kotlin/io/github/hellocuriosity/providers/StringProvider.kt
@@ -1,4 +1,4 @@
-package providers
+package io.github.hellocuriosity.providers
 
 import com.thedeanda.lorem.LoremIpsum
