+++ src/test/kotlin/io/github/hellocuriosity/ModelForgeTest.kt
@@ -1,3 +1,5 @@
+package io.github.hellocuriosity
+
 import org.junit.Test
 import java.util.Date
 import kotlin.random.Random
