+++ src/test/kotlin/io/github/hellocuriosity/providers/StringProviderTest.kt
@@ -1,4 +1,4 @@
-package providers
+package io.github.hellocuriosity.providers
 
 import org.junit.Test
 import org.junit.jupiter.api.Assertions.assertEquals
